+++ dao/game_result_dao.go
@@ -3,7 +3,6 @@ package dao
 import (
 	"context"
 	"errors"
-	"github.com/google/uuid"
 	"github.com/ildomm/cceab/domain"
 )
 
@@ -15,5 +14,5 @@ var ErrInvalidTransactionStatus = errors.New("invalid transaction status")
 
 type GameResultDAO interface {
 	CreateGameResult(ctx context.Context, gameResult domain.GameResult) (*domain.GameResult, error)
-	UpdateGameResult(ctx context.Context, gameResultId uuid.UUID, validationStatus domain.ValidationStatus) error
+	ValidateGameResults(ctx context.Context, totalGamesToCancel int) error
 }